extract settings
