+++ picasso/visualizations/base.py
@@ -1,3 +1,6 @@
+import re
+
+
 class BaseVisualization:
     """Interface encapsulating a NN visualization.
 
@@ -36,6 +39,27 @@ class BaseVisualization:
 
         """
         return self._model
+
+    def update_settings(self, settings):
+        def error_string(setting, setting_val):
+            return ('{val} is not an acceptable value for '
+                    'parameter {param} for visualization'
+                    '{vis}.').format(val=setting_val,
+                                     param=setting,
+                                     vis=self.__class__.__name__)
+
+        for setting in settings:
+            if settings[setting] in self.ALLOWED_SETTINGS[setting]:
+                # if the setting is allowed, set the attribute but remove
+                # invalid variable characters
+                #
+                # see:
+                #
+                # https://stackoverflow.com/questions/3303312/how-do-i-convert-a-string-to-a-valid-variable-name-in-python
+                setattr(self, re.sub('\W|^(?=\d)', '_', setting),
+                        settings[setting])
+            else:
+                raise ValueError(error_string(settings[setting], setting))
 
     def make_visualization(self, inputs, output_dir, settings=None):
         """Generate the visualization.
