+++ picasso/visualizations/partial_occlusion.py
@@ -33,14 +33,34 @@ class PartialOcclusion(BaseVisualization):
         'Occlusion': ['grey', 'black', 'white']
     }
 
+    @property
+    def window(self):
+        self._window = 0.10
+        return self._window
+
+    @window.setter
+    def window(self, val):
+        self._window = float(val)
+
+    @property
+    def num_windows(self):
+        self._strides = 20
+        return self._strides
+
+    @num_windows.setter
+    def num_windows(self, val):
+        return int(self._strides)
+
+    @property
+    def occlusion_method(self):
+        self._occlusion = 'white'
+        return self._occlusion
+
     def __init__(self, model):
         super().__init__(model)
         self.predict_tensor = self.get_predict_tensor()
 
-        self.window = 0.10
-        self.num_windows = 20
         self.grid_percent = 0.01
-        self.occlusion_method = 'white'
         self.occlusion_value = 255
         self.initial_resize = (244, 244)
 
@@ -100,33 +120,6 @@ class PartialOcclusion(BaseVisualization):
         # tensor in the computation graph
         return self.model.sess.graph.get_tensor_by_name(
             self.model.tf_predict_var.name)
-
-    def update_settings(self, settings):
-        def error_string(setting, setting_val):
-            return ('{val} is not an acceptable value for '
-                    'parameter {param} for visualization'
-                    '{vis}.').format(val=setting_val,
-                                     param=setting,
-                                     vis=self.__class__.__name__)
-
-        if 'Window' in settings:
-            if settings['Window'] in self.ALLOWED_SETTINGS['Window']:
-                self.window = float(settings['Window'])
-            else:
-                raise ValueError(error_string(settings['Window'], 'Window'))
-
-        if 'Strides' in settings:
-            if settings['Strides'] in self.ALLOWED_SETTINGS['Strides']:
-                self.num_windows = int(settings['Strides'])
-            else:
-                raise ValueError(error_string(settings['Strides'], 'Strides'))
-
-        if 'Occlusion' in settings:
-            if settings['Occlusion'] in self.ALLOWED_SETTINGS['Occlusion']:
-                self.occlusion_method = settings['Occlusion']
-            else:
-                raise ValueError(error_string(settings['Occlusion'],
-                                              'Occlusion'))
 
     def make_heatmaps(self, predictions,
                       output_dir, filename,
